fix: Extract absolute path to a system method
The tests depended on a certain absolute path being returned, which only
makes sense on one of my developer machines.
The main logic remains the same -- strictly a testing issue.
Going to try and remove the CI guard in the tests too -- as this is
hopefully the same problem I saw in GitHub actions.